requirements.txt
# Antigravity Core - Dependencies

# Core ML/Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Drone Control
pymavlink>=2.4.0

# Networking
pyzmq>=25.0.0

# Geospatial
shapely>=2.0.0
pyproj>=3.6.0

# Configuration
pyyaml>=6.0

# Image Processing (for geotagging)
Pillow>=10.0.0
piexif>=1.1.3

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: GUI/Visualization
# matplotlib>=3.7.0  # Uncomment for visualization
# folium>=0.14.0     # Uncomment for map display
